changed int to size_t

+++ subsequence.hpp
@@ -34,7 +34,7 @@ std::string sequence_to_string(const sequence& seq) {
 }
 
 // Generate a pseudorandom sequence of the given size, using the given
-// seed, where all elements are in the range [0, max_element]. 
+// seed, where all elements are in the range [0, max_element].
 // max_element must be non-negative.
 sequence random_sequence(size_t size, unsigned seed, int max_element) {
 
@@ -54,7 +54,7 @@ sequence random_sequence(size_t size, unsigned seed, int max_element) {
 
 bool is_nonincreasing(const sequence& A) {
     // if the next value is less than or equal to the current value for the whole vector, it is non-increasing
-    for(int i = 0; i < (A.size()-1); i++){
+    for(size_t i = 0; i < (A.size()-1); i++){
       if(A.at(i+1) > A.at(i)){ // if there is one case that the next value is greater than the current one, return false
         return false;
       }
@@ -63,45 +63,45 @@ bool is_nonincreasing(const sequence& A) {
   }
 
 sequence longest_nonincreasing_end_to_beginning(const sequence& A) {
-  
+
   const size_t n = A.size();
-  
+
   // populate the array H with 0 values
   std::vector<size_t> H(n, 0);
-  
+
   // calculate the values of array H
   // note that i has to be declared signed, to avoid an infinite loop, since
   // the loop condition is i >= 0
   for (signed int i = n-2;  i>= 0; i--) {
     for (size_t j = i+1; j < n ; j++) {
-      // checks the current value of A with all of the values that come before that index 
-      // also checks the current value of H with all of the values that come before that index 
+      // checks the current value of A with all of the values that come before that index
+      // also checks the current value of H with all of the values that come before that index
       if(A.at(i) >= A.at(j) && H.at(i) <= H.at(j)){
-        H.at(i) = H.at(j)+1; // add 1 to the current index's value 
+        H.at(i) = H.at(j)+1; // add 1 to the current index's value
       }
     }
   }
-  
+
   // calculate in max the length of the longest non-increasing subsequence
   // by adding 1 to the maximum value in H
   auto max = *std::max_element(H.begin(), H.end()) + 1;
-  
+
   // allocate space for the subsequence R
   std::vector<int> R(max);
-  
-  // add elements to R in non-increasing order whose H's values are 
+
+  // add elements to R in non-increasing order whose H's values are
   // in decreasing order, starting with the element whose H value is max-1
   // store in index the H values sought: max-1, max-2, max-3, .. 0
-  
+
   size_t index = max-1, j = 0;
   for (size_t i = 0; i < n; ++i) {
-    if (H[i] == index) { 
+    if (H[i] == index) {
       R.at(j) = (A.at(i)); // assign that specific value from vector A to vector R
-      index--; // decreasing index from max to 0 
+      index--; // decreasing index from max to 0
       j++;
     }
   }
-  
+
   return sequence(R.begin(), R.begin() + max);
 }
 
@@ -111,7 +111,7 @@ sequence longest_nonincreasing_powerset(const sequence& A) {
   std::vector<size_t> stack(n+1, 0);
   size_t k = 0;
   while (true) {
-    
+
     if (stack[k] < n) {
       stack[k+1] = stack[k] + 1;
       ++k;
@@ -119,11 +119,11 @@ sequence longest_nonincreasing_powerset(const sequence& A) {
       stack[k-1]++;
       k--;
     }
-    
+
     if (k == 0) {
       break;
     }
-    
+
     sequence candidate;
     // generate the candidate
     for (size_t i = 1; i <= k; ++i) {
